[WINUI/MESSUI] Added ATTR_PRINTF where necessary and fixed warnings
git-svn-id: svn://messdev.no-ip.org/mess@3883 9849079f-6229-0410-a07f-84406ce305ba

+++ src/osd/winui/directinput.c
@@ -100,7 +100,7 @@ BOOL DirectInputInitialize()
 
 		if (FAILED(hr))
 		{
-			ErrorMsg("DirectInputCreate failed! error=%x\n", hr);
+			ErrorMsg("DirectInputCreate failed! error=%x\n", (unsigned int)hr);
 			di = NULL;
 			return FALSE;
 		}
